fix: make htmlrewriter an external dependency
Previously, we accidentally bundled the htmlrewriter package into web-fragments repo.

Bundling the package in prevented the htmlrewriter resolutions to be honored within the client application.

See:
- https://github.com/anfibiacreativa/htmlrewriter/blob/fix-emoji-crash/package.json#L22
- https://github.com/anfibiacreativa/htmlrewriter/blob/fix-emoji-crash/cloudflare.js

Additionally by depending on the github fork we made it impossible to install the package in yarn v2 which has bugs in implementation of its github protocol.

This pr switches us over to temporarily published wf-htmlrewriter instead.