dev: refactor reframed by splitting it into several smaller modules
The goal here is to make it easier to follow and maintain this code.

I didn't go too crazy with the refactoring intentionally because I didn't want to the diff to be completely unreviewable.

It will likely take a few incremental refactorings to make this code significantly easier to follow.