fix: document.activeElement should return document.body if main document's activeElement is the main body
`<body>` is a special element in that it's often a fallback choice as the active element if no other element is active but the current document is active. Because of this we should treat the case when main body is the activeElement in a way that makes the fragment believe that it's own body is active.