dev: add eslint for playwright tests
Writing playwright tests without a linter is suicidal.

It's way way too easy to forget to add an  somewhere in test, which then results in the test running without waiting for the browser to process test instructions and start executing assertions before an action took place.

I'm not excited about bringing eslint into the project, but there is an existing playwright plugin for eslint, so I'm just going to use that.

The initial eslint configuration is very basic — most rules are disabled just to get the linter setup and lint passing (because of lack of checks).

References:
- https://playwright.dev/docs/best-practices#lint-your-tests
- https://www.npmjs.com/package/eslint-plugin-playwright