dev(jscftp): pnpm dev should support autoreload (autorestart)
I tried to implement this via vite-node's HMR but I couldn't figure out how to safely shut down the proxy server so that the proxy can reload, so I ended up using nodemon instead and restarting the whole vite-node process.

This is a bit slower but everything is so fast that it doesn't really matter.