docs: fix spacing issue on small screens
Previously, due to the way that the HTML was set up, there would be no
whitespace between the first two clauses of the "Start Now" section's
subtitle. This does not matter on larger screens, because there is a
`<br>` element that is included between them

This commit fixes this by ensuring that there is always a space after
the first clause, regardless of whether the `<br>` element is displayed
or not.

**Before the fix:**
![Web-fragments webpage - before][1]

**After the fix:**
![Weg-fragments webpage - after][2]

[1]: https://github.com/user-attachments/assets/40b55be4-2497-4053-9306-cd4edbd81266
[2]: https://github.com/user-attachments/assets/5000d358-3014-4271-ad00-0267c1081c9d